requirements.txt
fastapi
uvicorn
pydantic
zenml
torch
torchvision
scikit-learn
matplotlib
pillow
pymongo
